+++ apis.tf
@@ -1,6 +1,6 @@
 # Enable Cloud Asset API in billing export project
 resource "google_project_service" "cloud_asset" {
-  project = var.billing_account_id
+  project = var.billing_export_project_id
   service = "cloudasset.googleapis.com"
 
   disable_on_destroy         = var.disable_apis_on_destroy
@@ -9,7 +9,7 @@ resource "google_project_service" "cloud_asset" {
 
 # Enable Cloud Billing API in billing export project
 resource "google_project_service" "cloud_billing" {
-  project = var.billing_account_id
+  project = var.billing_export_project_id
   service = "cloudbilling.googleapis.com"
 
   disable_on_destroy         = var.disable_apis_on_destroy
@@ -19,7 +19,7 @@ resource "google_project_service" "cloud_billing" {
 # Enable Recommender API in billing export project
 # Note: Recommender API needs to be enabled per project, but recommendations are scoped to billing account
 resource "google_project_service" "recommender" {
-  project = var.billing_account_id
+  project = var.billing_export_project_id
   service = "recommender.googleapis.com"
 
   disable_on_destroy         = var.disable_apis_on_destroy
@@ -31,7 +31,7 @@ resource "google_project_service" "recommender" {
 # Most customers use on-demand pricing and can skip this
 resource "google_project_service" "bigquery_reservation" {
   count   = var.enable_bigquery_reservation_api ? 1 : 0
-  project = var.billing_account_id
+  project = var.billing_export_project_id
   service = "bigqueryreservation.googleapis.com"
 
   disable_on_destroy         = var.disable_apis_on_destroy